+++ samples/sampler/stories/default/PageViews.js
@@ -23,7 +23,12 @@ export default {
 };
 
 export const _PageViews = (args) => (
-	<PageViews fullContents={args['fullContents']} pageIndicatorPosition={args['pageIndicatorPosition']} pageIndicatorType={args['pageIndicatorType']}>
+	<PageViews
+		bannerMode={args['bannerMode']}
+		fullContents={args['fullContents']}
+		pageIndicatorPosition={args['pageIndicatorPosition']}
+		pageIndicatorType={args['pageIndicatorType']}
+	>
 		<PageViews.Page aria-label="This is a description for page 1">
 			<div style={{padding: '24px', width: '50%'}}>
 				<Item>Item 1</Item>
@@ -72,6 +77,7 @@ export const _PageViews = (args) => (
 );
 
 boolean('fullContents', _PageViews, Config, false);
+boolean('bannerMode', _PageViews, Config, false);
 select('pageIndicatorPosition', _PageViews, propOptions.pageIndicatorPosition, Config, 'bottom');
 select('pageIndicatorType', _PageViews, propOptions.pageIndicatorType, Config, 'dot');